use double instead of Float
git-svn-id: https://svn.salilab.org/imp/trunk@1929 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/algebra/include/Sphere3D.h
@@ -20,11 +20,11 @@ public:
   /** */
   Sphere3D(const Vector3D& center,double radius);
   /** */
-  Float get_volume() const;
+  double get_volume() const;
   /** */
-  Float get_surface_area() const;
+  double get_surface_area() const;
   /** */
-  Float get_radius() const {return radius_;}
+  double get_radius() const {return radius_;}
   /**  */
   const Vector3D &get_center() const {return center_;}
   /**  */
